global variables: a $ is prepended to each variable name
The file bin/gss-globalVariables collects pairs of the form (variable name) - value. Variable names
are scattered across several files of the piece. If a file has a reference to a global variable, this
reference should be of the form $VARIABLE_NAME, where VARIABLE_NAME is a variable from
bin/hss-globalVariables.

Variable names in bin/hss-globalVariables don't have a beginning $ sign.

Changes in
     bin/hss-globalVariables
     bin/hss-globalVariables
     bin/replaceGlobalVarNames.sh